+++ src/libs/libs/graph.h
@@ -8,7 +8,7 @@ typedef struct Graph Graph;
 Graph *graph_construct(int num_nodes);
 void graph_destruct(Graph *graph);
 int graph_get_num_nodes(Graph *graph);
-void graph_add_edge(Graph *graph, int src, int dest, long double cost);
+void graph_add_edge(Graph *graph, int src, int dest, double cost);
 void graph_read(Graph *graph, FILE *file, int num_edges);
 ForwardList *graph_get_edges(Graph *graph, int node);
